requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==2.3.7
email-validator==2.1.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
wtforms==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.6
fpdf2==2.7.6
pillow==10.1.0
gevent==23.9.1
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
